Read experiment flag(populate the link address lifetime) only once.
Although the experiment flag(IPCLIENT_POPULATE_LINK_ADDRESS_LIFETIM) is
read at the IpClient and IpClientLinkObserver constructor as a final
member, it's possbile that one code path will be updated but not the
other (if DeviceConfig is updated in the meantime). Instead of reading
the flag in IpClientLinkObserver again, put it to
IpClientLinkObserver#Configuration as a final member, refer to it when
it's necessary.

Bug: 323500923
Test: atest NetworkStackRootTests
(cherry picked from https://android-review.googlesource.com/q/commit:ee744e5f1d90b631086540d065f8d04a42917408)
Merged-In: I150dc167587484dbdaef5570542b7fa485437782
Change-Id: I150dc167587484dbdaef5570542b7fa485437782