Provide RRO configuration to customize preferred properties for filling DHCP client hostname option
Add a configurable option in the RRO which provides the customized preferred properties for filling DHCP client hostname option,
replacing the default device name (Dependent on config_dhcp_client_hostname is true).
If this value is set, the DHCP hostname option will be filled in with the value of the first property in the list that is not empty.
Otherwise, the DHCP hostname option will be filled in with the device name set in Settings.Global.DEVICE_NAME.
For example:
  <item>ro.product.model</item>
  <item>ro.product.name</item>

Bug: 343450973
Test: manual test, create empty APK to overlay the RRO configuration.
Change-Id: Ie55758b63209484e242352016888e69e3d288d7d