Reduce the performance impact of reading APF counters
This removes the counter read before every program installation which
can (and generally will) happen *very* frequently. Instead, increase the
polling frequency to 5 minutes. Overall, this should lead to a drastic
reduction of HAL calls.

It shall also be noted that this additional call to
startReadPacketFilter was never added to LegacyApfFilter, so this code
path only becomes active once the ApfFilter experiment rolls out.

Bug: 467472257
Test: TH
Change-Id: Ifc613e8fbd346bba476f642fed2dcb75500bdb35