Consistently use Flag annotations for mocking feature flags
Create consistent mock functionality for calls to both
isFeatureEnabled() and isFeatureNotChickenedOut(). Binary pushing a flag
(by replacing an instance of isFeatureEnabled() with its
isFeatureNotChickenedOut() counterpart) should not affect the tests. The
flag has yet to be removed, so the test should continue to test both
code paths.

This change defines both mocks for isFeatureEnabled() and
isFeatureNotChickenedOut() for all features defined in @Flag
annotations. It also cleans up a couple of preexising
isFeatureNotChickenedOut() mocks to be consistent with the rest of the
test class. Note that a missing @Flag annotation means the feature is
disabled (or chickened out).

Test: atest NetworkStackNextTests --test-filter .*IpReachabilityMonitorTest
Change-Id: I286bc41a08ce2395f2c2463a8d281c41c7852a48