Make ignore NUD failures integration tests can run against MTS.
In order to run NUD failure ignore relevant test cases against MTS, we
have to stop using the mocked IpMemoryStoreClient in the test, instead,
we can retrieve the specific event counts from db to verify if a new
network event (i.e. NUD failure) is written to db or not.

This CL implements the storeNetworkEvent API for signature test, it
simulates to write an event to the map. And also implement the assert
function to verify the retrieved network events.

Bug: 329010113
Test: atest NetworkStackSignatureTests
Change-Id: Ide540ca6add2ff1a477a2c95a1bb6dd328e79030