Wait for expected callback in DDR mode change test
The previous code used "verify" with the same arguments as the previous
call to "verify", so it did not actually wait for the second callback
that should be fired, and returned immediately. So the test is not
verifying what it intends to, and this is flaky as the second callback
could be received before and fail the "times(1)" check.

Bug: 368873634
Test: atest --rerun-until-failure 100
Change-Id: If017bf6e5c050f1294e5b1de7d1ba30bdcaa5b33