Rename sendPacketToUnreachableNeighbor to sendPacketToPeer.
sendPacketToUnReachableNeighbor is primarily used to trigger address
resolution for a neighbor and simulate the NUD failure by not responding
to the probes such as ARP or NS, then verify if the neighbor lost
callback is triggered or not. Due to the flag which ignores the NUD
failure from a neighbor which is never reachable has been enabled by
default, the test code might call this function to send a UDP packet to
a off-line peer or reachable neighbors and trigger address resolution
again, so rename it with a general purpose name.

Bug: 382194967
Test: atest NetworkStackRootTests
Change-Id: I61b4524ac521dee5e25d3d98a7db0b2acb468d8a