Update the prepareIpReachabilityMonitorTest behavior.
Previously prepareIpReachabilityMonitorTest intends not to respond to
the NS during the address resolution, it results in a NUD failure from
a neighbor which is not reachable and verifies the neighbor lost
callback. This verification doesn't work any more due to the flag which
ignores the NUD failure from a neighbor which is never reachable has
been enabled by default.

This CL updates the prepareIpReachabilityMonitorTest behavior by
responding to the NS during address resolution, which makes the neighbor
reachable. After that, simulate a L2 roam event, this will trigger
kernel to probe all neighbors again, but we don't respond to the probes
at this time, which causes another NUD failure post roam, it should
throw neighbor lost callback given the neighbor was ever reachable.

In this way we can remove the flag annotation from some test cases
which still require the flag with false value, i.e. the feature is
disabled by default, it doesn't match the actual production code.

Bug: 382194967
Test: atest NetworkStackRootTests
Test: atest NetworkStackIntegrationTests
Change-Id: I164b3a79b6da4a0755fce1acc643b26aa346cbad