Fix flaky testOnIpv6AddressRemoved test.
This test case was designed to delete all global IPv6 addresses from the
tap interface and then expect to see a provisioning loss event. The IPv6
addresses are retrieved from LinkProperties in order, the first global
IPv6 address to be deleted is the stable privacy address, then the
temporary privacy address follows.

However, sometimes the test case fails due to a ServiceSpecificException
with EADDRNOTAVAIL error is thrown when deleteing the temporary privacy
address. It's because that the temporary privacy address has been
deleted already along with deleting the stable privacy address, but the
test case still iterates the loop and tries to delete this temporary
privacy address again, so it ends up throwing an EADDRNOTAVAIL error.

To fix this flakiness, delete the temporary privacy address first and
then the stable privacy one, that won't introduce any race.

Bug: 384862724
Test: atest NetworkStackIntegrationTest --rerun-until-failure
Change-Id: I8b72662dc4f8e44d3fec06c19193764a979a72a9