Fail open on APF program generation failure.
When APF filter generation fails (due to program size is too large or
other errors), this commit make it fail open by zeroing the APF program
(but still keep the counter region unchanges) instead of abandoning the
update. Because this usually indicates a bug in the APF generation code
or an misconfigured system, failing open is the safer approach.

Test: TH
Change-Id: If850ec6297e172ffb671b5f1f9ee01a1115b334f