Add addJumpIfPktAtR0ContainDnsQ() API for multiple qTypes
This commit adds a new API, addJumpIfPktAtR0ContainDnsQ(), to the APF
generators to support matching a single qname with multiple qTypes. This
enhancement prepares the codebase for APFv6.1, which will introduce new
opcodes to improve code density by matching one qname with multiple
qTypes (e.g., SRV/TXT and A/AAAA records sharing the same qname in
mDNS). This commit focuses on creating the API structure; a follow-up
commit will update the Apf61GeneratorBase to utilize the new opcodes
with this API.

Test: TH
Change-Id: I29ff68030651f5a55e10d7ba365f9a595683aa77