Switch to the two-way version of INetworkMonitorCallbacks.
This is required to resolve the races in early NetworkAgent
creation.

This interface is called by NetworkMonitor and calls into
ConnectivityService. Because the NetworkMonitor code is in the
networkstack APK and ConnectivityService is in the tethering
APEX, the client can be newer than the server (e.g., R device, or
S device that does not support APEX). This is fine because per
the AIDL team, one-way vs two-way behaviour os called by the
client. So the newer client will force the server to return the
response (and will block until that happens).

Presubmit testing covers both combinations where both
networkstack and tethering have this CL, and where only
networkstack has this CL, so basic functionality is tested
in presubmit.

Additionally:
- On devices that do not have the tethering module (R devices,
  non-APEX devices) the ConnectivityService code that implements
  the callbacks could theoretically throw exceptions - either
  RemoteException due to method not being implemented, or any
  other type of parcelable exception (e.g., NullPointerException,
  IllegalArgumentException, etc.). The previous code would ignore
  those exceptions because the interface was marked as one-way,
  but now that the interface is two-way, these exceptions will be
  propagated to NetworkMonitor code and could crash it.
- Fix NetworkStackServiceTest#testDumpVersion. Now that the
  interface is two-way, the use of a spy breaks the test the
  caller will wait for the result and will get a RemoteException.
  Because Q is now unsupported, we can just use a mock like the
  test does for all other IPCs (INetworkMonitorCallbacks was
  special-cased because the AIDL interface had an argument that
  didn't exist on Q).

Test: TH
Bug: 296355968
Change-Id: Iba334e02569bb3ad714205d9d7ee6978a182d020