Add getRaProgramLengthOverEstimate() helper method.
This commit introduces the getRaProgramLengthOverEstimate() helper
method. This method allows us to estimate the program size for RA
filter without adding RA program to the actual APF program. This
simplifies size estimation and prepares for a future change to generate
the full program only once in installNewProgram(), instead of twice.

Test: TH
Change-Id: I36fe9de47f2b022ea63f913d9aab23a9a1446af5